+++ core-ng/src/main/java/core/framework/impl/template/Template.java
@@ -5,7 +5,6 @@ import core.framework.impl.template.fragment.CompositeFragment;
 import core.framework.impl.template.fragment.Fragment;
 import core.framework.impl.template.function.Function;
 import core.framework.impl.template.function.HTMLFunction;
-import core.framework.impl.template.function.IfFunction;
 
 import java.util.Map;
 
@@ -38,7 +37,6 @@ public class Template extends CompositeFragment {
 
     private void addBuiltInFunctions(CallStack stack) {
         stack.functions.put("html", new HTMLFunction());
-        stack.functions.put("if", new IfFunction());
     }
 
     @Override