+++ core-ng/src/main/java/core/framework/impl/template/location/StringTemplateSource.java
@@ -8,9 +8,11 @@ import java.io.StringReader;
  * @author neo
  */
 public final class StringTemplateSource implements TemplateSource {
+    public final String name;
     public final String template;
 
-    public StringTemplateSource(String template) {
+    public StringTemplateSource(String name, String template) {
+        this.name = name;
         this.template = template;
     }
 
@@ -22,5 +24,10 @@ public final class StringTemplateSource implements TemplateSource {
     @Override
     public TemplateSource resolve(String path) {
         throw new Error("string template does not support include");
+    }
+
+    @Override
+    public String source() {
+        return name;
     }
 }