+++ core-ng/src/main/java/core/framework/impl/template/location/TemplateSource.java
@@ -10,4 +10,6 @@ public interface TemplateSource {
     BufferedReader reader() throws IOException;
 
     TemplateSource resolve(String path);
+
+    String source();
 }