+++ demo-site/src/main/java/app/web/IndexPage.java
@@ -1,9 +1,21 @@
 package app.web;
 
+import core.framework.api.util.Lists;
+
+import java.util.List;
+
 /**
  * @author neo
  */
 public class IndexPage {
     public String name;
     public String backgroundColor;
+
+    public String welcomeMessage() {
+        return "hello " + name;
+    }
+
+    public List<String> items() {
+        return Lists.newArrayList("a", "b", "c");
+    }
 }